requirements.txt
fastapi
uvicorn
openai>=1.0.0
httpx
pydantic
python-dotenv
pageindex
